Several GWT compatibility fixes.
This allows gdx-ai to be used on GWT without the complex series of workarounds in #95 , because StandaloneFileSystem is swapped out for GdxFileSystem's implementation when running on GWT (which should always have libGDX available).

There are also some key fixes for reflection regarding behavior trees; GWT needs a lot of behavior tree classes made available in the class reflection pool.

Gradle is updated to 6.7.1, the same version libGDX currently uses to build. I want to update this to 7.3, but one step at a time.